made good progress using hunter and cmake directly.
I managed to get everything to compile and the tests pass.

The tricky part was getting the linking to agree with the includes.
it took me a while to include my Collections library.
The key was that the #include directive had to match the file path
exaclty (collections/linked_list.h)

The other key was in order to link against GTest I had to update the
hunter archive to 0.23.260

Next steps:
    run the test through valgrind and with coverage
    try including the library in a different project